Change walkers to use ints instead of floats to make manipulation easier.

+++ core/src/com/generator/main/generators/HullGenerator.java
@@ -9,7 +9,7 @@ public class HullGenerator {
     private int minArea;
     private float marginSize;
     private int targetArea;
-    private float sideLength;
+    private int sideLength;
     private int possibleStepsPerMove;
     private Random rand = new Random();
 
@@ -19,7 +19,7 @@ public class HullGenerator {
         this.marginSize = marginSize;
         this.possibleStepsPerMove = possibleStepsPerMove;
         targetArea = Math.round(minArea + (minArea * marginSize));
-        sideLength = (float) (Math.sqrt(targetArea));
+        sideLength = (int) (Math.sqrt(targetArea));
     }
 
     public Polygon generateSymmetricHull() throws Exception {
@@ -28,7 +28,7 @@ public class HullGenerator {
         Polygon output = new Polygon();
         ArrayList<Float> leftPath = new ArrayList<Float>();
         ArrayList<Float> rightPath = new ArrayList<Float>();
-        RandomWalker leftWalker = new RandomWalker(rand.nextFloat(0, sideLength/2),
+        RandomWalker leftWalker = new RandomWalker(rand.nextInt(0, sideLength/2),
                 0,
                 possibleStepsPerMove);
         leftPath.add(leftWalker.getXY()[0]);
@@ -38,7 +38,7 @@ public class HullGenerator {
         while (!(finished)){
             for (int i = 0; i<5; i++){
                 try {
-                    leftWalker.move(0.0F, sideLength/2);
+                    leftWalker.move(0, sideLength/2);
                     errorCount = 0;
                     leftPath.add(leftWalker.getXY()[0]);
                     leftPath.add(leftWalker.getXY()[1]);
